% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{dynverse}
\alias{dynverse}
\alias{dynverse-package}
\title{Easy installing and loading of the dynverse}
\description{
The 'dynverse' is a set of packages with which to
evaluate trajectory inference methods for dynamic processes.
}
